MaheswariKoyyani :Refractor the code to calculate Fare

+++ src/main/java/com/bridgelabs/cabpayrollservice/CabInvoice.java
@@ -5,7 +5,7 @@ import org.apache.logging.log4j.Logger;
 
 public class CabInvoice {
 	private static final Logger logger = LogManager.getLogger(CabInvoice.class);
-	public static final int MINIMUM_COST_PER_KILOMETER =10;
+	public static final double MINIMUM_COST_PER_KILOMETER =10.0;
 	public static final int COST_PER_TIME=1;
 
 	public static void main(String[] args) {
